+++ src/main/java/com/ssafy/todolist/service/EmailService.java
@@ -1,0 +1,49 @@
+package com.ssafy.todolist.service;
+
+import com.ssafy.todolist.domain.EmailResponse;
+import jakarta.mail.internet.MimeMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+@Service
+public class EmailService {
+
+    private final JavaMailSender javaMailSender;
+    private static int number;
+
+    public EmailService(JavaMailSender javaMailSender) {
+        this.javaMailSender = javaMailSender;
+    }
+
+    public MimeMessage createMail(String email) {
+        number = (int) (Math.random() * (90000)) + 100000;
+        MimeMessage message = javaMailSender.createMimeMessage();
+
+        try {
+            message.setFrom(email);
+            message.setRecipients(MimeMessage.RecipientType.TO, email);
+            message.setSubject("이메일 인증");
+            StringBuilder body = new StringBuilder();
+            body.append("<h3>요청하신 인증 번호 입니다.</h3>")
+                    .append("<h1>")
+                    .append(number)
+                    .append("</h1>")
+                    .append("<h3>감사합니다.</h3>");
+            message.setText(body.toString(), "UTF-8", "html");
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return message;
+    }
+
+    public EmailResponse sendCertificationNumber(String email) {
+        try {
+            javaMailSender.send(createMail(email));
+            return new EmailResponse(true);
+        } catch (Exception e) {
+            return new EmailResponse(false);
+        }
+    }
+
+}
